Fixed an issue where you could no longer build the project because of the newer version of g++; Cleaned up the code; Optimized preview picture for README

+++ src/backend/dominantcolor/DominantColor.h
@@ -3,11 +3,12 @@
 
 #include <memory>
 #include <vector>
+#include <string_view>
 
 namespace TermColor::Utils {
 
 struct color {
-    int r, g, b;
+    size_t r, g, b;
 
     const std::string getCommaSeparatedStr() const;
     const std::string getHexStr() const;
